fixed bugs on case loan
git-svn-id: http://nt-svn-jira.vdc.csc.com/HPF/Implementation/Source Code@2449 6d9d970a-e046-46de-96ce-fbc9f6c46dd9

+++ HPF.FutureState/HPF.FutureState.Web/InvoicePayments/ViewEditInvoicePaymentUC.ascx.cs
@@ -196,7 +196,7 @@ namespace HPF.FutureState.Web.InvoicePayments
             {
                 try
                 {
-                    int.Parse(txtPaymentAmt.Text);
+                    double.Parse(txtPaymentAmt.Text);
                 }
                 catch
                 {
